requirements.txt
spotipy~=2.17.1
fastapi~=0.63.0
PyYAML~=5.4.1